// src/types/IMarketDepth.tsx
export interface IMarketDepth {
  tk: string;
  e: string;
  name: string;
  bp: string;
  bp1: string;
  bp2: string;
  bp3: string;
  bp4: string;
  sp: string;
  sp1: string;
  sp2: string;
  sp3: string;
  sp4: string;
  bq: string;
  bq1: string;
  bq2: string;
  bq3: string;
  bq4: string;
  bs: string;
  bs1: string;
  bs2: string;
  bs3: string;
  bs4: string;
  bno1: string;
  bno2: string;
  bno3: string;
  bno4: string;
  bno5: string;
  sno1: string;
  sno2: string;
  sno3: string;
  sno4: string;
  sno5: string;
  mul: string;
  prec: string;
  ts: string;
}